# specs/002-backend-api-data/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Todo Task Management API
  description: |
    RESTful API for managing todo tasks in a multi-user application.
    All endpoints require JWT-based authentication.
    Users can only access and modify their own tasks.
  version: 1.0.0
  contact:
    name: Todo App API Support
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://api.example.com
    description: Production server (placeholder)

security:
  - BearerAuth: []

tags:
  - name: tasks
    description: Task CRUD operations

paths:
  /api/tasks:
    get:
      summary: List all tasks for authenticated user
      description: |
        Returns all tasks owned by the authenticated user, ordered by creation date (newest first).
        Tasks are automatically filtered by user_id from JWT token.
      tags:
        - tasks
      operationId: listTasks
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Successfully retrieved task list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
              examples:
                emptyList:
                  summary: User with no tasks
                  value: []
                multipleTasks:
                  summary: User with multiple tasks
                  value:
                    - id: "550e8400-e29b-41d4-a716-446655440001"
                      user_id: "123e4567-e89b-12d3-a456-426614174000"
                      title: "Buy groceries"
                      completed: false
                      created_at: "2026-02-05T14:30:00Z"
                      updated_at: "2026-02-05T14:30:00Z"
                    - id: "550e8400-e29b-41d4-a716-446655440002"
                      user_id: "123e4567-e89b-12d3-a456-426614174000"
                      title: "Call dentist"
                      completed: true
                      created_at: "2026-02-04T10:15:00Z"
                      updated_at: "2026-02-05T09:20:00Z"
        '401':
          $ref: '#/components/responses/UnauthorizedError'

    post:
      summary: Create a new task
      description: |
        Creates a new task for the authenticated user.
        The user_id is automatically set from the JWT token.
        The task is created with completed=false by default.
      tags:
        - tasks
      operationId: createTask
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
            examples:
              validTask:
                summary: Valid task creation
                value:
                  title: "Buy groceries"
              emojiTask:
                summary: Task with emoji
                value:
                  title: "Call mom 📞"
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
              example:
                id: "550e8400-e29b-41d4-a716-446655440003"
                user_id: "123e4567-e89b-12d3-a456-426614174000"
                title: "Buy groceries"
                completed: false
                created_at: "2026-02-05T15:00:00Z"
                updated_at: "2026-02-05T15:00:00Z"
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  /api/tasks/{task_id}:
    parameters:
      - name: task_id
        in: path
        required: true
        description: UUID of the task to operate on
        schema:
          type: string
          format: uuid
        example: "550e8400-e29b-41d4-a716-446655440001"

    get:
      summary: Get a single task
      description: |
        Retrieves a single task by ID.
        Returns 403 Forbidden if the task belongs to a different user.
      tags:
        - tasks
      operationId: getTask
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Task retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Invalid task ID format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                detail:
                  - loc: ["path", "task_id"]
                    msg: "value is not a valid uuid"
                    type: "type_error.uuid"
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'

    patch:
      summary: Toggle task completion status
      description: |
        Updates the completion status of a task.
        Returns 403 Forbidden if the task belongs to a different user.
        Automatically updates the updated_at timestamp.
      tags:
        - tasks
      operationId: toggleTaskCompletion
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskToggle'
            examples:
              markComplete:
                summary: Mark task as complete
                value:
                  completed: true
              markIncomplete:
                summary: Mark task as incomplete
                value:
                  completed: false
      responses:
        '200':
          description: Task completion status updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Invalid task ID format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'

    put:
      summary: Update task title
      description: |
        Updates the title of an existing task.
        Returns 403 Forbidden if the task belongs to a different user.
        Automatically updates the updated_at timestamp.
      tags:
        - tasks
      operationId: updateTaskTitle
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
            example:
              title: "Buy groceries and cook dinner"
      responses:
        '200':
          description: Task title updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'

    delete:
      summary: Delete a task
      description: |
        Permanently deletes a task.
        Returns 403 Forbidden if the task belongs to a different user.
      tags:
        - tasks
      operationId: deleteTask
      security:
        - BearerAuth: []
      responses:
        '204':
          description: Task deleted successfully (no content)
        '400':
          description: Invalid task ID format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token issued by Better Auth during signin.
        Include in Authorization header as: `Bearer <token>`

  schemas:
    Task:
      type: object
      required:
        - id
        - user_id
        - title
        - completed
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the task
          example: "550e8400-e29b-41d4-a716-446655440001"
        user_id:
          type: string
          format: uuid
          description: ID of the user who owns this task
          example: "123e4567-e89b-12d3-a456-426614174000"
        title:
          type: string
          minLength: 1
          maxLength: 500
          description: Task title or description
          example: "Buy groceries"
        completed:
          type: boolean
          description: Whether the task is completed
          example: false
        created_at:
          type: string
          format: date-time
          description: Timestamp when the task was created (UTC)
          example: "2026-02-05T14:30:00Z"
        updated_at:
          type: string
          format: date-time
          description: Timestamp when the task was last updated (UTC)
          example: "2026-02-05T14:30:00Z"

    TaskCreate:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 500
          description: Task title (cannot be empty or whitespace-only)
          example: "Buy groceries"

    TaskUpdate:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 500
          description: New task title (cannot be empty or whitespace-only)
          example: "Buy groceries and cook dinner"

    TaskToggle:
      type: object
      required:
        - completed
      properties:
        completed:
          type: boolean
          description: New completion status
          example: true

    Error:
      type: object
      properties:
        detail:
          oneOf:
            - type: string
              description: Simple error message
              example: "Task not found"
            - type: array
              description: Validation error details
              items:
                type: object
                properties:
                  loc:
                    type: array
                    items:
                      oneOf:
                        - type: string
                        - type: integer
                    description: Location of the error (e.g., ["body", "title"])
                  msg:
                    type: string
                    description: Error message
                  type:
                    type: string
                    description: Error type
              example:
                - loc: ["body", "title"]
                  msg: "ensure this value has at most 500 characters"
                  type: "value_error.any_str.max_length"

  responses:
    UnauthorizedError:
      description: Missing, invalid, or expired JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: "Could not validate credentials"

    ForbiddenError:
      description: Valid token but attempting to access another user's resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: "Access forbidden"

    NotFoundError:
      description: Task with the specified ID does not exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: "Task not found"

    ValidationError:
      description: Request validation failed (empty title, too long, etc.)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            emptyTitle:
              summary: Empty or whitespace-only title
              value:
                detail: "Task title cannot be empty or whitespace"
            titleTooLong:
              summary: Title exceeds 500 characters
              value:
                detail:
                  - loc: ["body", "title"]
                    msg: "ensure this value has at most 500 characters"
                    type: "value_error.any_str.max_length"